change decimals to integer

+++ backend/api/models.py
@@ -398,26 +398,26 @@ class ProjectsSalesResults(models.Model):
         max_length=10, primary_key=True , editable=False
     )
     project = models.ForeignKey(Projects, on_delete=models.CASCADE, null=True)
-    sales_revenue = models.DecimalField(max_digits=12, decimal_places=2, default=0.0)
-    dispatch_labor_expense = models.DecimalField(
+    sales_revenue = models.IntegerField(max_digits=12, decimal_places=2, default=0.0)
+    dispatch_labor_expense = models.IntegerField(
         max_digits=12, decimal_places=2, default=0.0
     )
-    employee_expense = models.DecimalField(
+    employee_expense = models.IntegerField(
         max_digits=12, decimal_places=2, default=0.0
     )
-    indirect_employee_expense = models.DecimalField(
+    indirect_employee_expense = models.IntegerField(
         max_digits=12, decimal_places=2, default=0.0
     )
-    expense = models.DecimalField(max_digits=12, decimal_places=2, default=0.0)
-    operating_income = models.DecimalField(max_digits=12, decimal_places=2, default=0.0)
-    non_operating_income = models.DecimalField(
+    expense = models.IntegerField(max_digits=12, decimal_places=2, default=0.0)
+    operating_income = models.IntegerField(max_digits=12, decimal_places=2, default=0.0)
+    non_operating_income = models.IntegerField(
         max_digits=12, decimal_places=2, default=0.0
     )
-    non_operating_expense = models.DecimalField(
+    non_operating_expense = models.IntegerField(
         max_digits=12, decimal_places=2, default=0.0
     )
-    ordinary_profit = models.DecimalField(max_digits=12, decimal_places=2, default=0.0)
-    ordinary_profit_margin = models.FloatField(default=0.0)
+    ordinary_profit = models.IntegerField(max_digits=12, decimal_places=2, default=0.0)
+    ordinary_profit_margin = models.IntegerField(default=0.0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     class Meta :
